Add support for multiple channels.
This is not visible anyway. Switching to a different channel is done by the "/join roomname" command. A notice is displayed, telling that the user left/joined, but on indication on the channel name is giving. You have to coordinate before switching channels, otherwise nobody knows where you are.

The default channel is called "Lobby", that's where everybody starts at.